functions/requirements.txt
firebase-functions
firebase-admin
google-generativeai>=0.8.3
flask
cors